+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## 2.7.0-SNAPSHOT
+- Added WorldGuard region blacklist
+  - You can add certain WorldGuard regions to your `disabled-worldguard-regions` list in the config.yml. Players in that region will not spawn an AngelChest.
 - Added Russian translation
 - Removed xp settings from groups.default.yml because that is controlled via permissions